greeting the user + removed redundent class closes #47
