+++ src/main/java/Team5/Fitnesspeaker/AlexaCommunication/Handlers/LaunchRequestHandler.java
@@ -27,7 +27,8 @@ public class LaunchRequestHandler implements RequestHandler {
 
 	@Override
 	public Optional<Response> handle(final HandlerInput i) {
-		final String speechText = "Welcome to the fitnesspeaker, for more information, please say help";
+		final String name = i.getServiceClientFactory().getUpsService().getProfileGivenName();
+		final String speechText = "Hey "+name+", Welcome to nutracker! for more information, please say help";
 		final String repromptText = "I will repeat, " + speechText;
 		return i.getResponseBuilder().withSimpleCard("FitnessSpeakerSession", speechText).withSpeech(speechText)
 				.withReprompt(repromptText).build();
